requirements.txt
streamlit
Pillow
diffusers
torch
transformers
